Everywhere: Enough to build and run js/test-js
Fails on GC for some reason, can be forced with adding -g to
line 192 of `ladybird.py`

Most of this is brought over from my WindowsJS2.0 branch however
we have regressed since then with gc. My old fix for SymbolOrString
doesnt work anymmore so ill have to investigate it further.